requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==2.0.4
django-bower==5.2.0
djangorestframework==3.8.2
idna==2.6
lxml==4.2.1
pytz==2018.4
requests==2.18.4
six==1.11.0
urllib3==1.22
